perf: save space by reducing size if id BREAKING CHANGE: use integer as type of ids to save space

+++ server/src/main/java/com/backend/common/JwtData.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class JwtData {
-    private Long userId;
+    private Integer userId;
     private String userName;
-    private List<Long> roleIds;
+    private List<Integer> roleIds;
     private List<String> roleNames;
 }
