+++ server/src/main/java/com/backend/service/dto/PagedResponse.java
@@ -11,7 +11,7 @@ public class PagedResponse<T> {
     private List<T> content;
     private int page;
     private int size;
-    private long totalElements;
+    private Integer totalElements;
     private int totalPages;
     private boolean last;
 }
